backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
PyJWT
sqlparse
paho.mqtt
daphne
channels
channels-redis